fix(carvers): adversarial-review findings + Task 8 verification record
24-agent review (6 dimensions vs A1-A7, javap tiebreak, adversarial
verify per finding): zero shipped-data parity deviations. Fixed the
confirmed datapack/UB-discipline items:

- seed+carverIndex reseed: unsigned add (Java ladd wraps; signed UB for
  seeds within 2 of INT64_MAX — own UBSan gate would fire)
- ellipsoid column disc gate: dcmpl+iflt polarity — Java PROCESSES the
  column on NaN; !(sum<1.0) skipped it (A2 gloss annotated)
- canyon distance: jvm_f2i saturating cast (NaN->0, overflow->INT_MAX/MIN)
- debug_settings.debug_mode now fail-loud (was silently ignored; gates
  mask/replaceable bypass + debug-state substitution in vanilla)
- float providers: explicit minecraft:constant form accepted; vanilla
  codec validations added (probability floatRange(0,1), uniform max>min
  strict, trapezoid max>=min and plateau<=span)
- hc_lcg_next_int: debug assert bound>0 (Java throws; pow2 path returned
  0 silently)
- json.c strtod comment scoped: bit-match holds for Double.parseDouble
  consumers only; Codec.FLOAT double-rounding gap deferred to Task 12

VERIFICATION.md: mutation-probe matrix (7 covered / 9 blind, all blind
verdicts mechanistically explained), canyon-disable experiment (canyon
contributes 0 cells to the gate on this seed), full review record,
Task 9/12 handoff. Gates re-run post-fix: 18/18 ctest, no-FMA (17,344
insns), ASan+UBSan full suite PASS.

+++ core/src/gen_carvers_stage.c
@@ -50,9 +50,12 @@ void hc_gen_carvers_stage(hc_chunk_t *chunk, hc_noise_chunk_t *nc,
             for (int32_t idx = 0; idx < n_carvers; idx++) {
                 /* 인덱스는 isStartChunk 결과와 무관하게 증가; 시드
                  * 항은 (worldSeed + idx) 이고 청크 좌표는 믹스에서
-                 * 들어간다 (A1 §4.1/§9) */
-                hc_lcg_set_large_feature_seed(&rng, seed + (int64_t)idx, scx,
-                                              scz);
+                 * 들어간다 (A1 §4.1/§9). ladd 는 mod 2^64 랩 —
+                 * INT64_MAX 근방 시드에서 signed UB 가 되지 않도록
+                 * 무부호로 더한다 (리뷰 확정: verify:orchestration). */
+                hc_lcg_set_large_feature_seed(
+                    &rng, (int64_t)((uint64_t)seed + (uint64_t)idx), scx,
+                    scz);
                 env.cv = &carvers[idx];
                 /* isStartChunk: nextFloat() <= probability, 포함 비교
                  * (A3 §2 / A4 §2) */
